+++ lib/Fhp/Message/Message.php
@@ -104,7 +104,7 @@ class Message extends AbstractMessage
         $this->systemId = $systemId;
         $this->options = $options;
         $this->profileVersion = SecurityProfile::PROFILE_VERSION_2;
-        $this->securityFunction = HNSHK::SECURITY_FUNC_900;
+        $this->securityFunction = HNSHK::SECURITY_FUNC_999;
 
         if (isset($options[static::OPT_PINTAN_MECH])) {
             if (!in_array('999', $this->options[static::OPT_PINTAN_MECH])) {
